logic changes
