+++ .docker/nginx/conf.d/default.conf
@@ -5,6 +5,8 @@ server {
 
     index index.php index.html index.htm;
 
+    client_max_body_size 200M;  
+
     location / {
         try_files $uri $uri/ =404;
     }
