+++ .docker/php/php.ini
@@ -1,0 +1,2 @@
+upload_max_filesize = 200M
+post_max_size = 200M
