+++ src/Models/PartnerModel.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Framework\Models;
+
+use Framework\Repository\PartnerRepository;
+
+class PartnerModel extends Model
+{
+    protected static string $repository = PartnerRepository::class;
+}
