+++ src/Repository/PartnerRepository.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Framework\Repository;
+
+class PartnerRepository extends Repository
+{
+    protected static string $table = "partners";
+}
