+++ src/Repository/ProjectRepository.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Framework\Repository;
+
+class ProjectRepository extends Repository
+{
+    protected static string $table = "projects";
+}
