--- src/framework/Base/Models/ProjectModel.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Framework\Models;
-
-use Framework\Repository\ProjectRepository;
-
-class ProjectModel extends Model
-{
-    protected static string $repository = ProjectRepository::class;
-}
