--- src/framework/Base/Repository/PartnerRepository.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Framework\Repository;
-
-class PartnerRepository extends Repository
-{
-    protected static string $table = "partners";
-}
