--- src/framework/Base/Repository/ProjectRepository.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Framework\Repository;
-
-class ProjectRepository extends Repository
-{
-    protected static string $table = "projects";
-}
