--- src/framework/Base/Services/FileWriter.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Framework\Services;
-
-class FileWriter 
-{
-    
-}
