--- src/framework/framework.php
@@ -1,10 +1,0 @@
-<?php
-
-
-//dd($_SERVER["DOCUMENT_ROOT"]);
-
-
-global $APPLICATION;
-
-//$APPLICATION = new CMain();
-
