--- src/logs.log
@@ -1,329 +1,0 @@
-PAGEN=1
-PAGEN=2
-Error in parsPartnerBlock: Could not load file https://www.1c-bitrix.ru/partners/212563.php.Error in parsPartnerBlock: Could not load file https://www.1c-bitrix.ru/partners/151690.php.Error in parsPartnerBlock: Could not load file https://www.1c-bitrix.ru/partners/1420148.php.Error in getPartnerAddress: Call to a member function getAttribute() on null
-PAGEN=3
-Could not load file https://www.1c-bitrix.ru/partners/index_ajax.php?PAGEN_1=4.
-Could not load file https://www.1c-bitrix.ru/partners/index_ajax.php?PAGEN_1=4.
-PAGEN=4
-PAGEN=5
-PAGEN=6
-PAGEN=7
-PAGEN=8
-PAGEN=9
-Could not load file https://www.1c-bitrix.ru/partners/index_ajax.php?PAGEN_1=10.
-PAGEN=10
-PAGEN=11
-PAGEN=12
-PAGEN=13
-PAGEN=14
-PAGEN=15
-PAGEN=16
-PAGEN=17
-PAGEN=18
-PAGEN=19
-PAGEN=20
-PAGEN=21
-PAGEN=22
-Error in getPartnerName: Call to a member function text() on null
-Error in getPartnerAddress: Call to a member function getAttribute() on null
-Error in parsPartnerBlock: Could not load file https://www.1c-bitrix.ru/partners/384269.php.Error in parsPartnerBlock: Could not load file https://www.1c-bitrix.ru/partners/10412700.php.Error in parsPartnerBlock: Could not load file https://www.1c-bitrix.ru/partners/1397032.php.PAGEN=23
-PAGEN=24
-PAGEN=25
-PAGEN=26
-PAGEN=27
-PAGEN=28
-PAGEN=29
-PAGEN=30
-PAGEN=31
-PAGEN=32
-PAGEN=33
-PAGEN=34
-PAGEN=35
-PAGEN=36
-PAGEN=37
-PAGEN=38
-PAGEN=39
-PAGEN=40
-Error in getPartnerName: Call to a member function text() on null
-Error in getPartnerAddress: Call to a member function getAttribute() on null
-PAGEN=41
-Error in parsPartnerBlock: Could not load file https://www.1c-bitrix.ru/partners/11430870.php.PAGEN=42
-PAGEN=43
-PAGEN=44
-PAGEN=45
-PAGEN=46
-PAGEN=47
-PAGEN=48
-Error in getPartnerName: Call to a member function text() on null
-Error in getPartnerAddress: Call to a member function getAttribute() on null
-PAGEN=49
-PAGEN=50
-PAGEN=51
-PAGEN=52
-PAGEN=53
-PAGEN=54
-PAGEN=55
-Error in getPartnerName: Call to a member function text() on null
-Error in getPartnerAddress: Call to a member function getAttribute() on null
-PAGEN=1
-Error in getPartnerName: Call to a member function text() on null
-Error in getPartnerAddress: Call to a member function getAttribute() on null
-PAGEN=2
-PAGEN=3
-PAGEN=4
-PAGEN=5
-Error in getPartnerName: Call to a member function text() on null
-Error in getPartnerAddress: Call to a member function getAttribute() on null
-PAGEN=1
-Error in getPartnerName: Call to a member function text() on null
-Error in getPartnerAddress: Call to a member function getAttribute() on null
-PAGEN=6
-PAGEN=2
-PAGEN=7
-PAGEN=3
-PAGEN=8
-PAGEN=4
-PAGEN=9
-PAGEN=5
-PAGEN=10
-PAGEN=6
-PAGEN=11
-Error in getPartnerName: Call to a member function text() on null
-Error in getPartnerAddress: Call to a member function getAttribute() on null
-PAGEN=7
-PAGEN=12
-PAGEN=8
-PAGEN=13
-PAGEN=9
-PAGEN=14
-PAGEN=10
-PAGEN=15
-PAGEN=11
-Error in getPartnerName: Call to a member function text() on null
-Error in getPartnerAddress: Call to a member function getAttribute() on null
-PAGEN=16
-PAGEN=12
-PAGEN=17
-PAGEN=13
-PAGEN=18
-PAGEN=14
-PAGEN=19
-PAGEN=15
-PAGEN=20
-PAGEN=16
-PAGEN=21
-PAGEN=17
-PAGEN=22
-PAGEN=18
-PAGEN=23
-PAGEN=19
-PAGEN=24
-PAGEN=20
-PAGEN=25
-PAGEN=21
-PAGEN=26
-PAGEN=22
-PAGEN=27
-PAGEN=23
-PAGEN=28
-PAGEN=24
-PAGEN=25
-PAGEN=29
-PAGEN=26
-PAGEN=30
-PAGEN=27
-PAGEN=31
-PAGEN=28
-PAGEN=32
-PAGEN=29
-PAGEN=33
-PAGEN=30
-PAGEN=34
-PAGEN=31
-PAGEN=35
-PAGEN=32
-PAGEN=36
-PAGEN=33
-PAGEN=37
-PAGEN=34
-PAGEN=38
-PAGEN=35
-PAGEN=39
-PAGEN=36
-PAGEN=40
-PAGEN=37
-PAGEN=41
-PAGEN=38
-PAGEN=42
-PAGEN=39
-PAGEN=43
-PAGEN=40
-Error in getPartnerName: Call to a member function text() on null
-Error in getPartnerAddress: Call to a member function getAttribute() on null
-PAGEN=44
-PAGEN=41
-PAGEN=45
-PAGEN=42
-PAGEN=46
-PAGEN=43
-Error in getPartnerName: Call to a member function text() on null
-Error in getPartnerAddress: Call to a member function getAttribute() on null
-PAGEN=44
-PAGEN=47
-PAGEN=45
-PAGEN=48
-PAGEN=46
-PAGEN=49
-PAGEN=47
-PAGEN=50
-PAGEN=48
-PAGEN=51
-PAGEN=49
-PAGEN=52
-PAGEN=50
-PAGEN=53
-PAGEN=51
-PAGEN=54
-PAGEN=52
-PAGEN=55
-PAGEN=53
-PAGEN=56
-PAGEN=54
-PAGEN=57
-PAGEN=55
-PAGEN=58
-PAGEN=56
-PAGEN=57
-PAGEN=59
-PAGEN=58
-PAGEN=60
-PAGEN=59
-PAGEN=61
-PAGEN=60
-PAGEN=62
-PAGEN=61
-PAGEN=63
-PAGEN=62
-PAGEN=64
-PAGEN=63
-PAGEN=65
-PAGEN=64
-PAGEN=66
-PAGEN=65
-Error in parsPartnerBlock: Could not load file https://www.1c-bitrix.ru/partners/175253.php.Error in parsPartnerBlock: Could not load file https://www.1c-bitrix.ru/partners/278785.php.PAGEN=67
-PAGEN=66
-PAGEN=68
-PAGEN=67
-PAGEN=69
-PAGEN=68
-PAGEN=70
-PAGEN=69
-PAGEN=71
-PAGEN=70
-PAGEN=72
-PAGEN=71
-PAGEN=73
-PAGEN=72
-PAGEN=73
-PAGEN=74
-PAGEN=74
-PAGEN=75
-PAGEN=75
-Error in getPartnerName: Call to a member function text() on null
-Error in getPartnerAddress: Call to a member function getAttribute() on null
-PAGEN=76
-Error in getPartnerName: Call to a member function text() on null
-Error in getPartnerAddress: Call to a member function getAttribute() on null
-PAGEN=76
-PAGEN=77
-PAGEN=77
-PAGEN=78
-PAGEN=78
-PAGEN=79
-PAGEN=79
-PAGEN=80
-PAGEN=80
-PAGEN=81
-PAGEN=81
-PAGEN=82
-PAGEN=82
-PAGEN=83
-PAGEN=83
-PAGEN=84
-PAGEN=84
-PAGEN=85
-PAGEN=85
-PAGEN=86
-PAGEN=86
-PAGEN=87
-PAGEN=87
-PAGEN=88
-PAGEN=88
-PAGEN=89
-PAGEN=89
-PAGEN=90
-PAGEN=90
-PAGEN=91
-PAGEN=91
-PAGEN=92
-PAGEN=92
-PAGEN=93
-PAGEN=93
-PAGEN=94
-PAGEN=94
-PAGEN=95
-PAGEN=95
-PAGEN=96
-PAGEN=96
-PAGEN=97
-PAGEN=97
-PAGEN=98
-PAGEN=98
-PAGEN=99
-PAGEN=99
-PAGEN=100
-PAGEN=100
-PAGEN=101
-PAGEN=101
-PAGEN=102
-PAGEN=102
-PAGEN=103
-PAGEN=103
-PAGEN=104
-PAGEN=104
-PAGEN=105
-PAGEN=105
-PAGEN=106
-PAGEN=106
-PAGEN=107
-PAGEN=107
-PAGEN=108
-PAGEN=108
-PAGEN=109
-PAGEN=109
-PAGEN=110
-PAGEN=110
-PAGEN=111
-PAGEN=111
-PAGEN=112
-PAGEN=112
-PAGEN=113
-PAGEN=113
-PAGEN=114
-PAGEN=114
-PAGEN=115
-PAGEN=115
-PAGEN=116
-PAGEN=116
-PAGEN=117
-PAGEN=117
-PAGEN=118
-PAGEN=118
-PAGEN=119
-PAGEN=119
-Error in getPartnerAddress: Call to a member function getAttribute() on null
-Error in getPartnerAddress: Call to a member function getAttribute() on null
-Error in getPartnerAddress: Call to a member function getAttribute() on null
-Error in getPartnerAddress: Call to a member function getAttribute() on null
-PAGEN=120
-PAGEN=120
